1. 修复mode4发送123指令工作期间设备会离线的问题

+++ H01_Air780E_app/Task/app_port.c
@@ -533,7 +533,7 @@ __attribute__((interrupt("WCH-Interrupt-fast")))
 __attribute__((section(".highcode")))
 void GPIOA_IRQHandler(void)
 {
-    uint32_t iqr;
+    uint16_t iqr;
     iqr = GPIOA_ReadITFlagPort();
     if (iqr & RING_PIN)
     {
@@ -573,7 +573,7 @@ __attribute__((interrupt("WCH-Interrupt-fast")))
 __attribute__((section(".highcode")))
 void GPIOB_IRQHandler(void)
 {
-    uint32_t iqr;
+    uint16_t iqr;
     iqr = GPIOB_ReadITFlagPort();
     if (iqr & GSINT_PIN)
     {
